feat: implement ContentView and MovieCard with MVVM binding
