+++ FigmaToSwiftUIApp/FigmaToSwiftUIAppApp.swift
@@ -6,12 +6,22 @@
 //
 
 import SwiftUI
+import Models
+import ViewModels
 
 @main
 struct FigmaToSwiftUIAppApp: App {
+    // Replace with your actual TMDb API key
+    private let movieService = MovieService(apiKey: "YOUR_API_KEY")
+    private let movieListViewModel: MovieListViewModel
+
+    init() {
+        movieListViewModel = MovieListViewModel(movieService: movieService)
+    }
+
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            ContentView(viewModel: movieListViewModel)
         }
     }
 }
