+++ FigmaToSwiftUIApp/Models/Entities/Movie.swift
@@ -1,0 +1,46 @@
+import Foundation
+
+struct Movie: Identifiable, Codable {
+    let id: Int
+    let title: String
+    let overview: String
+    let posterPath: String
+    let releaseDate: String
+    let voteAverage: Double
+
+    // MARK: - Coding Keys
+    enum CodingKeys: String, CodingKey {
+        case id
+        case title
+        case overview
+        case posterPath = "poster_path"
+        case releaseDate = "release_date"
+        case voteAverage = "vote_average"
+    }
+
+    // MARK: - Computed Properties
+
+    /// Formatted release date in "MMM d, yyyy" format
+    var formattedReleaseDate: String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "yyyy-MM-dd"
+
+        guard let date = dateFormatter.date(from: releaseDate) else {
+            return "Release date unavailable"
+        }
+
+        dateFormatter.dateFormat = "MMM d, yyyy"
+        return dateFormatter.string(from: date)
+    }
+
+    /// Formatted rating as a percentage
+    var rating: String {
+        let percentage = Int(voteAverage * 10)
+        return "\(percentage)%"
+    }
+
+    /// Full poster URL
+    var posterURL: URL? {
+        URL(string: "https://image.tmdb.org/t/p/w500\(posterPath)")
+    }
+}
