+++ FigmaToSwiftUIApp/Models/Entities/MovieResponse.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+struct MovieResponse: Codable {
+    let page: Int
+    let totalPages: Int
+    let results: [Movie]
+
+    enum CodingKeys: String, CodingKey {
+        case page
+        case totalPages = "total_pages"
+        case results
+    }
+}
